+++ CHANGELOG.md
@@ -16,6 +16,9 @@ This is the 4th release candidate, a *pre-beta* version.
    `fieldBackground.param`, `particleDefinition.param`
    see the lines that should be removed in
    [#234](https://github.com/ComputationalRadiationPhysics/picongpu/pull/234/files)
+ - Renamed `observer.hpp` -> `radiationObserver.param` #237
+   Changed variable name `N_theta` to `N_observer`
+   [https://github.com/PrometheusPi/picongpu/commit/9e487ec30ade10ece44fc19fd7a815b8dfe58f61#diff-9]
 
 **New Features:**
  - n/a